# docs/swagger.yaml
basePath: /
definitions:
...
  domain.GetOrder:
    properties:
      address:
        type: string
      cost:
        type: number
      date:
        type: string
      id:
        type: string
      orderId:
        type: integer
      status:
        type: string
    type: object
...
  domain.OrderByID:
    properties:
      address:
        type: string
      clientFullName:
        type: string
      clientPhoneNumber:
        type: string
      cost:
        type: number
      courierService:
        type: integer
      deliveryTime:
        type: string
      deliveryType:
        type: string
      description:
        type: string
      dishes:
        items:
          $ref: '#/definitions/domain.OrdersDish'
        type: array
      feedback_id:
        type: string
      id:
        type: string
      orderId:
        type: integer
      paymentType:
        type: string
      rating:
        type: integer
      status:
        type: string
    type: object
...
  domain.OrdersDish:
    properties:
      amount:
        type: integer
      name:
        type: string
    type: object
...
  domain.UpdateOrder:
    properties:
      courierService:
        type: integer
      deliveryType:
        type: integer
      status:
        type: integer
    type: object
...
  handler.CheckNewOrdersMark:
    properties:
      data:
        type: boolean
    type: object
...
  handler.GetAllOrders:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.GetOrder'
        type: array
    type: object
  handler.GetOrderByID:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.OrderByID'
        type: array
    type: object
  handler.dataResponse:
    properties:
      data: {}
    type: object
...
  handler.response:
    properties:
      message:
        type: string
    type: object
...
host: 165.232.68.67:8080
info:
  contact: {}
  description: REST API for Restaurant-Assistant service
  title: Restaurant-Assistant API
  version: "1.0"
paths:
  /authenticated/order/check-new-orders-mark:
    get:
      consumes:
      - application/json
      description: Get status of presence of active orders
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.CheckNewOrdersMark'
      security:
      - manager-auth: []
      tags:
      - order
...
  /order:
    get:
      consumes:
      - application/json
      description: Get orders with filter, ordered by created time
      parameters:
      - description: select orders with created date > start
        format: date
        in: query
        name: start
        type: string
      - description: select orders with created date < end
        format: date
        in: query
        name: end
        type: string
      - description: restaurant ID
        in: query
        name: restaurantID
        type: string
      - description: dish name
        in: query
        name: dishName
        type: string
      - description: order status
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        in: query
        name: status
        type: integer
      - default: 20
        description: number of returned orders
        in: query
        name: limit
        type: integer
      - default: 1
        description: page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.GetAllOrders'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.dataResponse'
      tags:
      - order
  /order/{id}:
    get:
      consumes:
      - application/json
      description: get order information
      parameters:
      - description: order id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.GetOrderByID'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - order
    put:
      consumes:
      - application/json
      description: update order status
      parameters:
      - description: order update info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateOrder'
      - description: order id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - order
...
securityDefinitions:
  ManagerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"